Splitting pagination tests
