+++ Octokit.GraphQL.Core.UnitTests/PagingTests/Models/CommentModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Octokit.GraphQL.Core.UnitTests.PagingTests.Models
+{
+    class CommentModel
+    {
+        public string Id { get; set; }
+        public string Body { get; set; }
+    }
+}
