+++ Octokit.GraphQL.Core.UnitTests/PagingTests/Models/RepositoryModel.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Octokit.GraphQL.Core.UnitTests.PagingTests.Models
+{
+    class RepositoryModel
+    {
+        public string Name { get; set; }
+        public IList<IssueModel> Issues { get; set; }
+    }
+}
