+++ Octokit.GraphQL.Core.UnitTests/PagingTests/Models/RepositoryModelWithDictionary.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Octokit.GraphQL.Core.UnitTests.PagingTests.Models
+{
+    class RepositoryModelWithDictionary
+    {
+        public string Name { get; set; }
+        public IDictionary<int, IssueModel> Issues { get; set; }
+    }
+}
